+++ src/CScsi_Error_Counter_Log.cpp
@@ -2,7 +2,7 @@
 // CScsi_Error_Counter_Log.cpp  Definition of Error Counter for READ WRITE VERIFY ERRORS
 // Do NOT modify or remove this copyright and license
 //
-// Copyright (c) 2015 - 2018 Seagate Technology LLC and/or its Affiliates, All Rights Reserved
+// Copyright (c) 2014 - 2020 Seagate Technology LLC and/or its Affiliates, All Rights Reserved
 //
 // This software is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -312,7 +312,7 @@ void CScsiErrorCounterLog::process_Error_Data(JSONNODE *errorData)
             else
             {
                 std::string printStr;
-                snprintf((char*)printStr.c_str(), BASIC, "%" PRIu32"", m_ErrorValue);
+                snprintf((char*)printStr.c_str(), BASIC, "%" PRIu32"",  static_cast<uint32_t>(m_ErrorValue));
                 json_push_back(errorInfo, json_new_a((char *)myStr.c_str(), (char*)printStr.c_str()));
             }
         }
