+++ src/CScsi_Self_Test_Results_Log.cpp
@@ -2,7 +2,7 @@
 // CScsi_Self_Test_Results_Log.cpp  Implementation of DST Results log 
 // Do NOT modify or remove this copyright and license
 //
-// Copyright (c) 2015 - 2018 Seagate Technology LLC and/or its Affiliates, All Rights Reserved
+// Copyright (c) 2014 - 2020 Seagate Technology LLC and/or its Affiliates, All Rights Reserved
 //
 // This software is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -186,9 +186,9 @@ void CScsi_DST_Results::print_Self_Test_Log(JSONNODE *dstNode, uint16_t run)
 	snprintf((char*)myStr.c_str(), BASIC, "0x%02" PRIx8"", m_DST->paramControlByte);
 	json_push_back(runInfo, json_new_a("Control Byte", (char*)myStr.c_str()));
 
-	snprintf((char*)myStr.c_str(), BASIC, "0x%02" PRIx8"", M_GETBITRANGE(m_DST->stCode, 7, 5));
+	snprintf((char*)myStr.c_str(), BASIC, "0x%02" PRIx8"",(uint8_t) M_GETBITRANGE(m_DST->stCode, 7, 5));
 	json_push_back(runInfo, json_new_a("Self Test Code", (char*)myStr.c_str()));
-	snprintf((char*)myStr.c_str(), BASIC, "0x%02" PRIx8"", M_GETBITRANGE( m_DST->stCode,3,0));
+	snprintf((char*)myStr.c_str(), BASIC, "0x%02" PRIx8"", (uint8_t)M_GETBITRANGE( m_DST->stCode,3,0));
 	json_push_back(runInfo, json_new_a("Self Test Results", (char*)myStr.c_str()));
 	if (M_GETBITRANGE(m_DST->stCode, 7, 5) == DST_NOT_RUN)
 	{
